requirements.txt
asgiref==3.6.0
Django==4.2
gunicorn==20.1.0
Pillow==9.5.0
sorl-thumbnail==12.9.0
sqlparse==0.4.3
tzdata==2023.3